+++ packages/rangelink-vscode-extension/src/__tests__/helpers/index.ts
@@ -1,3 +1,4 @@
+export * from './captureSendStrategies';
 export * from './configureEmptyTabGroups';
 export * from './configureWorkspaceMocks';
 export * from './createBaseMockPasteDestination';