+++ packages/rangelink-vscode-extension/src/feedback/types/index.ts
@@ -1,4 +1,4 @@
-export * from './AutoUnbindReason';
+export * from './AutoUnbindDetails';
 export * from './DestinationInfo';
 export * from './PasteContext';
 export * from './PasteSendOutcome';